+++ src/storage/storage_async.rs
@@ -4,16 +4,21 @@ use async_recursion::async_recursion;
 
 use futures::{StreamExt, TryStreamExt};
 use itertools::Itertools;
+use serde::{Deserialize, Serialize};
 
 use crate::{
+    array::{ArrayMetadata, ArrayMetadataV2, ArrayMetadataV3},
     byte_range::ByteRange,
-    group::GroupMetadataV3,
+    group::{GroupMetadata, GroupMetadataV3},
+    metadata::GroupMetadataV2,
     node::{Node, NodeMetadata, NodePath},
+    storage::meta_key_v3,
 };
 
 use super::{
-    data_key, meta_key, AsyncBytes, MaybeAsyncBytes, StorageError, StoreKey, StoreKeyRange,
-    StoreKeyStartValue, StoreKeys, StoreKeysPrefixes, StorePrefix, StorePrefixes,
+    data_key, meta_key_v2_array, meta_key_v2_attributes, meta_key_v2_group, AsyncBytes,
+    MaybeAsyncBytes, StorageError, StoreKey, StoreKeyRange, StoreKeyStartValue, StoreKeys,
+    StoreKeysPrefixes, StorePrefix, StorePrefixes,
 };
 
 /// Async readable storage traits.
@@ -302,6 +307,77 @@ impl<T> AsyncReadableWritableListableStorageTraits for T where
 {
 }
 
+async fn get_metadata_v3<
+    TStorage: ?Sized + AsyncReadableStorageTraits + AsyncListableStorageTraits,
+>(
+    storage: &Arc<TStorage>,
+    prefix: &StorePrefix,
+) -> Result<Option<NodeMetadata>, StorageError> {
+    #[derive(Serialize, Deserialize)]
+    #[serde(untagged)]
+    enum NodeMetadataV3 {
+        Array(ArrayMetadataV3),
+        Group(GroupMetadataV3),
+    }
+
+    let key: StoreKey = meta_key_v3(&prefix.try_into()?);
+    match storage.get(&key).await? {
+        Some(metadata) => {
+            let metadata: NodeMetadataV3 = serde_json::from_slice(&metadata)
+                .map_err(|err| StorageError::InvalidMetadata(key, err.to_string()))?;
+            Ok(Some(match metadata {
+                NodeMetadataV3::Array(array) => NodeMetadata::Array(ArrayMetadata::V3(array)),
+                NodeMetadataV3::Group(group) => NodeMetadata::Group(GroupMetadata::V3(group)),
+            }))
+        }
+        None => Ok(None),
+    }
+}
+
+async fn get_metadata_v2<
+    TStorage: ?Sized + AsyncReadableStorageTraits + AsyncListableStorageTraits,
+>(
+    storage: &Arc<TStorage>,
+    prefix: &StorePrefix,
+) -> Result<Option<NodeMetadata>, StorageError> {
+    let node_path = prefix.try_into()?;
+    let attributes_key = meta_key_v2_attributes(&node_path);
+
+    // Try array
+    let key_array: StoreKey = meta_key_v2_array(&node_path);
+    if let Some(metadata) = storage.get(&key_array).await? {
+        let mut metadata: ArrayMetadataV2 = serde_json::from_slice(&metadata)
+            .map_err(|err| StorageError::InvalidMetadata(key_array, err.to_string()))?;
+        let attributes = storage.get(&attributes_key).await?;
+        if let Some(attributes) = attributes {
+            let attributes: serde_json::Map<String, serde_json::Value> =
+                serde_json::from_slice(&attributes).map_err(|err| {
+                    StorageError::InvalidMetadata(attributes_key, err.to_string())
+                })?;
+            metadata.attributes = attributes;
+        }
+        return Ok(Some(NodeMetadata::Array(ArrayMetadata::V2(metadata))));
+    }
+
+    // Try group
+    let key_group: StoreKey = meta_key_v2_group(&node_path);
+    if let Some(metadata) = storage.get(&key_group).await? {
+        let mut metadata: GroupMetadataV2 = serde_json::from_slice(&metadata)
+            .map_err(|err| StorageError::InvalidMetadata(key_group, err.to_string()))?;
+        let attributes = storage.get(&attributes_key).await?;
+        if let Some(attributes) = attributes {
+            let attributes: serde_json::Map<String, serde_json::Value> =
+                serde_json::from_slice(&attributes).map_err(|err| {
+                    StorageError::InvalidMetadata(attributes_key, err.to_string())
+                })?;
+            metadata.attributes = attributes;
+        }
+        return Ok(Some(NodeMetadata::Group(GroupMetadata::V2(metadata))));
+    }
+
+    Ok(None)
+}
+
 /// Asynchronously get the child nodes.
 ///
 /// # Errors
@@ -318,15 +394,14 @@ where
     let mut nodes: Vec<Node> = Vec::new();
     // TODO: Asynchronously get metadata of all prefixes
     for prefix in &prefixes {
-        let key = meta_key(&prefix.try_into()?);
-        let child_metadata = match storage.get(&key).await? {
-            Some(child_metadata) => {
-                let metadata: NodeMetadata = serde_json::from_slice(&child_metadata)
-                    .map_err(|err| StorageError::InvalidMetadata(key, err.to_string()))?;
-                metadata
-            }
-            None => NodeMetadata::Group(GroupMetadataV3::default().into()),
+        let mut child_metadata = get_metadata_v3(storage, prefix).await?;
+        if child_metadata.is_none() {
+            child_metadata = get_metadata_v2(storage, prefix).await?;
+        }
+        let Some(child_metadata) = child_metadata else {
+            return Err(StorageError::MissingMetadata(prefix.clone()));
         };
+
         let path: NodePath = prefix.try_into()?;
         let children = match child_metadata {
             NodeMetadata::Array(_) => Vec::default(),
@@ -414,10 +489,9 @@ pub async fn async_node_exists<
     storage: &Arc<TStorage>,
     path: &NodePath,
 ) -> Result<bool, StorageError> {
-    Ok(storage
-        .get(&meta_key(path))
-        .await
-        .map_or(storage.list_dir(&path.try_into()?).await.is_ok(), |_| true))
+    Ok(storage.get(&meta_key_v3(path)).await?.is_some()
+        || storage.get(&meta_key_v2_array(path)).await?.is_some()
+        || storage.get(&meta_key_v2_group(path)).await?.is_some())
 }
 
 /// Asynchronously check if a node exists.
@@ -429,12 +503,9 @@ pub async fn async_node_exists_listable<TStorage: ?Sized + AsyncListableStorageT
     path: &NodePath,
 ) -> Result<bool, StorageError> {
     let prefix: StorePrefix = path.try_into()?;
-    let parent = prefix.parent();
-    if let Some(parent) = parent {
-        storage.list_dir(&parent).await.map(|keys_prefixes| {
-            !keys_prefixes.keys().is_empty() || !keys_prefixes.prefixes().is_empty()
-        })
-    } else {
-        Ok(false)
-    }
+    storage.list_prefix(&prefix).await.map(|keys| {
+        keys.contains(&meta_key_v3(path))
+            | keys.contains(&meta_key_v2_array(path))
+            | keys.contains(&meta_key_v2_group(path))
+    })
 }